+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+- [windows][linux] optionally show the flag of the current IP country as the system tray icon
+
 # 1.5.0
 - [all] redesign ui with a modern dark theme